requirements.txt
streamlit
pandas
requests
weaviate-client>=3.26.7,<4.0.0
python-dotenv
langchain-ollama
langchain-core
pydantic
streamlit